requirements.txt
alembic==1.10.4
astor==0.8.1
blinker==1.6.2
boto3==1.26.133
botocore==1.29.133
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
common==0.1.2
data==0.4
decorator==5.1.1
dual==0.1.0
dynamo3==0.4.10
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask==2.3.2
flywheel==0.5.4
funcsigs==1.0.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
opt-einsum==3.3.0
paddle-bfloat==0.1.7
paddle-python==0.6.0
paddle==1.0.2
paddlepaddle==2.4.2
peewee==3.16.2
Pillow==9.5.0
pip==22.3.1
protobuf==3.20.0
prox==0.0.17
PyJWT==2.7.0
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-geoip-python3==1.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
setuptools==65.5.1
six==1.16.0
SQLAlchemy==2.0.13
tight==0.1.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
wheel==0.38.4
